Dropping old / obsolete / unmaintained code
Taking ownership in WP4 for the test_O2Device.cxx test
and renaming in to test_FairMQ.
