+++ Framework/Core/test/test_FairMQ.cxx
@@ -12,17 +12,104 @@
 #define BOOST_TEST_MAIN
 #define BOOST_TEST_DYN_LINK
 
-#include "O2Device/O2Device.h"
 #include "Headers/NameHeader.h"
+
+#include "MemoryResources/MemoryResources.h"
+#include "Headers/DataHeader.h"
+#include "Headers/Stack.h"
+
 #include <boost/test/unit_test.hpp>
 #include <iostream>
 #include <vector>
 #include <fairmq/Tools.h>
 #include <fairmq/ProgOptions.h>
+#include <gsl/gsl>
 
-using namespace o2::base;
 using namespace o2::header;
 using namespace o2::pmr;
+
+//__________________________________________________________________________________________________
+// addDataBlock for generic (compatible) containers, that is contiguous containers using the pmr allocator
+template <typename ContainerT, typename std::enable_if<!std::is_same<ContainerT, FairMQMessagePtr>::value, int>::type = 0>
+bool addDataBlock(FairMQParts& parts, o2::header::Stack&& inputStack, ContainerT&& inputData, o2::pmr::FairMQMemoryResource* targetResource = nullptr)
+{
+  using std::forward;
+  using std::move;
+
+  auto headerMessage = o2::pmr::getMessage(move(inputStack), targetResource);
+  auto dataMessage = o2::pmr::getMessage(forward<ContainerT>(inputData), targetResource);
+
+  parts.AddPart(move(headerMessage));
+  parts.AddPart(move(dataMessage));
+
+  return true;
+}
+
+//__________________________________________________________________________________________________
+// addDataBlock for data already wrapped in FairMQMessagePtr
+// note: since we cannot partially specialize function templates, use SFINAE here instead
+template <typename ContainerT, typename std::enable_if<std::is_same<ContainerT, FairMQMessagePtr>::value, int>::type = 0>
+bool addDataBlock(FairMQParts& parts, o2::header::Stack&& inputStack, ContainerT&& dataMessage, o2::pmr::FairMQMemoryResource* targetResource = nullptr)
+{
+  using std::move;
+
+  //make sure the payload size in DataHeader corresponds to message size
+  using o2::header::DataHeader;
+  DataHeader* dataHeader = const_cast<DataHeader*>(o2::header::get<DataHeader*>(inputStack.data()));
+  dataHeader->payloadSize = dataMessage->GetSize();
+
+  auto headerMessage = o2::pmr::getMessage(move(inputStack), targetResource);
+
+  parts.AddPart(move(headerMessage));
+  parts.AddPart(move(dataMessage));
+
+  return true;
+}
+
+template <typename I, typename F>
+auto forEach(I begin, I end, F&& function)
+{
+
+  using span = gsl::span<const o2::byte>;
+#ifdef MS_GSL_V3
+  using SPAN_SIZE_TYPE = span::size_type;
+#else
+  using SPAN_SIZE_TYPE = span::index_type;
+#endif
+  using gsl::narrow_cast;
+  for (auto it = begin; it != end; ++it) {
+    o2::byte* headerBuffer{nullptr};
+    SPAN_SIZE_TYPE headerBufferSize{0};
+    if (*it != nullptr) {
+      headerBuffer = reinterpret_cast<o2::byte*>((*it)->GetData());
+      headerBufferSize = narrow_cast<SPAN_SIZE_TYPE>((*it)->GetSize());
+    }
+    ++it;
+    o2::byte* dataBuffer{nullptr};
+    SPAN_SIZE_TYPE dataBufferSize{0};
+    if (*it != nullptr) {
+      dataBuffer = reinterpret_cast<o2::byte*>((*it)->GetData());
+      dataBufferSize = narrow_cast<SPAN_SIZE_TYPE>((*it)->GetSize());
+    }
+
+    // call the user provided function
+    function(span{headerBuffer, headerBufferSize}, span{dataBuffer, dataBufferSize});
+  }
+  return std::move(function);
+}
+
+/// Execute user code (e.g. a lambda) on each data block (header-payload pair)
+/// returns the function (same as std::for_each)
+template <typename F>
+auto forEach(FairMQParts& parts, F&& function)
+{
+  if ((parts.Size() % 2) != 0) {
+    throw std::invalid_argument(
+      "number of parts in message not even (n%2 != 0), cannot be considered an O2 compliant message");
+  }
+
+  return forEach(parts.begin(), parts.end(), std::forward<F>(function));
+}
 
 BOOST_AUTO_TEST_CASE(getMessage_Stack)
 {
@@ -80,7 +167,7 @@ BOOST_AUTO_TEST_CASE(addDataBlockForEach_test)
 
   {
     //simple addition of a data block from an exisiting message
-    O2Message message;
+    FairMQParts message;
     auto simpleMessage = factoryZMQ->CreateMessage(10);
     addDataBlock(message,
                  Stack{allocZMQ, DataHeader{gDataDescriptionInvalid, gDataOriginInvalid, DataHeader::SubSpecificationType{0}}},
@@ -95,7 +182,7 @@ BOOST_AUTO_TEST_CASE(addDataBlockForEach_test)
       int j;
     };
     using namespace boost::container::pmr;
-    O2Message message;
+    FairMQParts message;
     std::vector<elem, polymorphic_allocator<elem>> vec(polymorphic_allocator<elem>{allocZMQ});
     vec.reserve(100);
     vec.push_back({1, 2});
