--- Utilities/O2Device/CMakeLists.txt
@@ -1,21 +1,0 @@
-# Copyright CERN and copyright holders of ALICE O2. This software is distributed
-# under the terms of the GNU General Public License v3 (GPL Version 3), copied
-# verbatim in the file "COPYING".
-#
-# See http://alice-o2.web.cern.ch/license for full licensing information.
-#
-# In applying this license CERN does not waive the privileges and immunities
-# granted to it by virtue of its status as an Intergovernmental Organization or
-# submit itself to any jurisdiction.
-
-o2_add_library(Device
-               SOURCES src/O2Device.cxx
-               PUBLIC_LINK_LIBRARIES FairMQ::FairMQ O2::MemoryResources
-                                     O2::Headers ms_gsl::ms_gsl
-                                     AliceO2::Monitoring)
-if(NOT APPLE)
-  o2_add_test(O2Device 
-              SOURCES test/test_O2Device.cxx
-              PUBLIC_LINK_LIBRARIES O2::Device
-              COMPONENT_NAME Device)
-endif()
