--- Utilities/O2MessageMonitor/test/runO2MessageMonitor.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-#here's how to use it:
-#o2-message-monitor --id source1 --mq-config O2MessageMonitor.json
-#o2-message-monitor --id sink1 --mq-config O2MessageMonitor.json
