--- Utilities/aliceHLTwrapper/.gitignore
@@ -1,1 +1,0 @@
-*~
