--- Utilities/aliceHLTwrapper/nodelist.sh
@@ -1,10 +1,0 @@
-#! /bin/bash
-
-# this script defines the nodelist as a subsidiary to AliceO2
-# launch scripts
-
-############## example ##############################
-nodelist=
-nodelist=(${nodelist[@]} localhost)
-#nodelist=(${nodelist[@]} someothernode)
-
